+++ Configula/Patches/ConfigurationManagerPatch.cs
@@ -3,14 +3,18 @@
 using System.Collections.Generic;
 using System.Reflection.Emit;
 
+using ConfigurationManager;
+
 using HarmonyLib;
+
+using UnityEngine;
 
 using static PluginConfig;
 
-[HarmonyPatch(typeof(ConfigurationManager.ConfigurationManager))]
+[HarmonyPatch(typeof(ConfigurationManager))]
 static class ConfigurationManagerPatch {
   [HarmonyTranspiler]
-  [HarmonyPatch(nameof(ConfigurationManager.ConfigurationManager.CalculateWindowRect))]
+  [HarmonyPatch(nameof(ConfigurationManager.CalculateWindowRect))]
   static IEnumerable<CodeInstruction> CalculateWindowRectTranspiler(IEnumerable<CodeInstruction> instructions) {
     return new CodeMatcher(instructions)
         .MatchForward(
@@ -19,7 +23,9 @@ static class ConfigurationManagerPatch {
             new CodeMatch(OpCodes.Ldc_I4))
         .ThrowIfInvalid("Could not patch CalculateWindowRect() width value!")
         .Advance(offset: 2)
-        .InsertAndAdvance(Transpilers.EmitDelegate(GetWidthDelegate))
+        .InsertAndAdvance(
+            new CodeInstruction(
+                OpCodes.Call, AccessTools.Method(typeof(ConfigurationManagerPatch), nameof(GetWidthDelegate))))
         .MatchForward(
             useEnd: false,
             new CodeMatch(OpCodes.Call, AccessTools.Method(typeof(UnityEngine.Screen), "get_height")),
@@ -27,7 +33,9 @@ static class ConfigurationManagerPatch {
             new CodeMatch(OpCodes.Sub))
         .ThrowIfInvalid("Could not patch CalculateWindowRect() height value!")
         .Advance(offset: 2)
-        .InsertAndAdvance(Transpilers.EmitDelegate(GetHeightDelegate))
+        .InsertAndAdvance(
+            new CodeInstruction(
+                OpCodes.Call, AccessTools.Method(typeof(ConfigurationManagerPatch), nameof(GetHeightDelegate))))
         .InstructionEnumeration();
   }
 
@@ -37,5 +45,24 @@ static class ConfigurationManagerPatch {
 
   static int GetHeightDelegate(int heightOffset) {
     return IsModEnabled.Value ? WindowHeightOffset.Value : heightOffset;
+  }
+
+  [HarmonyPrefix]
+  [HarmonyPatch(nameof(ConfigurationManager.DrawTooltip))]
+  static bool DrawTooltipPrefix(Rect area) {
+    if (IsModEnabled.Value && UseAlternateDrawTooltip.Value) {
+      ConfigulaManager.DrawTooltip(area);
+      return false;
+    }
+
+    return true;
+  }
+
+  [HarmonyPrefix]
+  [HarmonyPatch(nameof(ConfigurationManager.OnGUI))]
+  static void OnGUIPrefix() {
+    if (IsModEnabled.Value && Event.current.type == EventType.Repaint) {
+      GUI.tooltip = string.Empty;
+    }
   }
 }